requirements/base.txt
aiohttp-jinja2==1.2.0
aiohttp-oauth2==0.0.3
aiohttp-remotes==0.1.2
aiohttp==3.6.2
aioredis==1.3.1
alembic==1.4.2
apscheduler==3.6.3
asyncpgsa==0.26.3
gunicorn==20.0.4
marshmallow==3.6.1
sentry-sdk==0.16.0
psycopg2-binary==2.8.5
pyyaml==5.3.1
git+https://github.com/pyslackers/slack-sansio.git@2312217b72b6d3827ab6107ac20fdb05dc4cab6c#egg=slack-sansio
uvloop==0.14.0
